### 2021-11-09 #484 @dylan-turner25
rfema:  access data through the FEMA API.
**Reviewers:**
Submitting Author Name: Dylan Turner
Submitting Author Github Handle: @dylan-turner25
Repository:  https://github.com/dylan-turner25/rfema
Version submitted: 0.0.0.9000
Submission type: Standard
Version accepted: TBD
---

```
Package: rfema
Title: Access the openFEMA API
Version: 0.0.0.9000
    person(given = "Dylan",
           family = "Turner",
           email = "dylan.turner3@outlook.com",
Description: `rfema` allows users to access The Federal Emergency Management Agency's (FEMA) publicly available data through their API. The package provides a set of functions to easily navigate and access data from the National Flood Insurance Program along with FEMA's various disaster aid programs, including the Hazard Mitigation Grant Program, the Public Assistance Grant Program, and the Individual Assistance Grant Program.
    httr,
    memoise,
    plyr,
    rvest,
    utils
    covr,
    vcr (>= 0.6.0)

```

`rfema` allows users to access The Federal Emergency Management Agency's (FEMA) publicly available data through the open FEMA API. The package provides a set of functions to easily navigate and access all data sets provided by FEMA, including (but not limited to) data from the National Flood Insurance Program and FEMA's various disaster aid programs.

`rfema` is intended to be used by researchers that require the use of data that is available through the FEMA API, but may not have familiarity in working with APIs.

I am unaware of any other packages that offer similar functionality.

N/A

- [ x] I have read the [guide for authors](https://devguide.ropensci.org/guide-for-authors.html) and [rOpenSci packaging guide](https://devguide.ropensci.org/building.html).
