+++ CONTRIBUTING.md
@@ -1,0 +1,10 @@
+# Contributing
+
+If you plan to contribute code that would amount to a substantial change in the package, please first outline the planned change as an [issue](https://github.com/dylan-turner25/rfema/issues) so that it can be discussed to get community feedback.
+
+Contributors are asked to comply with the [code of conduct](https://github.com/dylan-turner25/rfema/CONDUCT.md) if they wish to remain associated with the project.
+
+# Pull Requests
+
+Pull requests are more than welcome. `rfema` generally follows the style laid out in Hadley Wickham and Jennifer Bryan's [R Packages](https://r-pkgs.org/). Conforming to the same development practices will mean review of your pull request will be faster and smoother. 
+
