### 2021-11-10 #484 @dylan-turner25
The most recent commit has attempted to address all of the failed initial checks.
