### 2021-11-11 beb8d83
fixing example in bulk_dl() function documentation.
