### 2021-11-12 #484 @emilyriederer
Hi @dylan-turner25 ! 👋 It's nice to meet you, and I'm excited to be working with you on this package. I'm going to start seeking out good reviewers for this package. In the meantime, I wanted to direct your attention to a few items from the automated checks:

- `lintr` suggests you have a lot of lines >80 characters long. For readability, it might be helpful to shorten. You can run [`lintr::lint_package()`](https://github.com/jimhester/lintr) locally in RStudio and R will point you to specific lines to reconsider. This doesn't have to be at zero, but if there are opportunities to add spacing and improve readability, that would be great.
- Consider adding some documentation to your one internal function (it looks like there may be none now). You can suppress `.Rd` generation with the `#' @noRd` tag , but it will help reviewers understand the code
- Please try to remove the one note being generated by R CMD check. The [Imports section](https://r-pkgs.org/namespace.html?q=import#imports) of the *R Packages* book discusses a few good options for defining your imports. In short, I think there's inconsistencies in what packages you are importing (via the `NAMESPACE`), which packages you are acknowledging in the `roxygen` docs (with `import` or `importFrom`), and how you are referencing those functions inside your code (with or without `::`)

Thanks again !
