### 2021-11-15 f42a4d3
Addressing issues raised in intial editor comments during Ropensci submission.
