+++ R/helpers.R
@@ -1,16 +1,22 @@
-#' Helper function to insure data_set parameter matches an open FEMA data offering.
+#' Helper function to insure data_set parameter matches an open FEMA data
+#'  offering.
 #'
-#' The function corrects any capitalization inconsistencies so that the "data_set" parameter in any of rfema's functions is not case sensative.
+#' The function corrects any capitalization inconsistencies so that the
+#' "data_set" parameter in any of rfema's functions is not case sensative.
 #'
 #' @param data_set a character string representing the data set.
 #'
-#' @return returns the data with capitalization changed to be consistent with FEMA's naming convention or returns an error if the data set is not one of the valid offerings through the FEMA API.
+#' @return returns the data with capitalization changed to be consistent with
+#' FEMA's naming convention or returns an error if the data set is not one of
+#' the valid offerings through the FEMA API.
 #' @export
+#' 
+#' @importFrom memoise memoise
 #'
 #' @examples
 #' valid_dataset("fimanfipclaims")
 #' valid_dataset("fIMANfipclaiMS")
-valid_dataset <- memoise(function(data_set) {
+valid_dataset <- memoise::memoise(function(data_set) {
 
 
   # get df with info on fema data sets
@@ -24,7 +30,10 @@ valid_dataset <- memoise(function(data_set) {
   # capitalization fema uses
   match <- fema_data_sets$name[tolower(fema_data_sets$name) == data_set]
   if (length(match) == 0) {
-    stop(paste0(data_set, " is not a valid data set. Use fema_data_sets() to return avaliable data sets and associated meta data."))
+    stop(paste0(
+      data_set,
+      "is not a valid data set. Use fema_data_sets() to return avaliable data sets and associated meta data."
+    ))
   } else {
     data_set <- match
   }
@@ -39,11 +48,19 @@ valid_dataset <- memoise(function(data_set) {
 #' Helper function to generate the URL string for an API query
 #'
 #' @param data_set character string specifiying the data set to query from
-#' @param top_n numeric value indicating how many records to retun (default is 1000 which is the maximum number of records the API will return in a single query )
-#' @param filters a list containing values of data fields that should be used to construct filters.
-#' @param select a character vector of data fields to return (default is to return all data fields for the data set)
+#' @param top_n numeric value indicating how many records to retun
+#' (default is 1000 which is the maximum number of records the API will return
+#' in a single query )
+#' @param filters a list containing values of data fields that should be used
+#' to construct filters.
+#' @param select a character vector of data fields to return
+#' (default is to return all data fields for the data set)
 #'
-#' @return The function returns a string containing a url that can be used to query the API.
+#' @return The function returns a string containing a url that can be used to
+#'  query the API.
+#'  
+#' @importFrom memoise memoise
+#' 
 #' @export
 #'
 #' @examples
@@ -53,7 +70,8 @@ valid_dataset <- memoise(function(data_set) {
 #'   data_set = "fimaNfipPolicies", top_n = 100,
 #'   filters = filter_list, select = vars_to_select
 #' )
-gen_api_query <- memoise(function(data_set = NULL, top_n = NULL, filters = NULL, select = NULL) {
+gen_api_query <- memoise::memoise(function(data_set = NULL, top_n = NULL, filters = NULL,
+                                  select = NULL) {
 
   # replace top_n with 1000 if no value is supplied
   if (is.null(top_n)) {
@@ -73,7 +91,10 @@ gen_api_query <- memoise(function(data_set = NULL, top_n = NULL, filters = NULL,
[…]
