+++ R/parameter_values.R
@@ -4,12 +4,13 @@
 #' @param data_field a character vector specifying the data field
 #'
 #' @return returns a data frame with the unique data field values conatained in the first 1000 records of the full data set
-#' @export
+#' 
 #' @importFrom memoise memoise
+#' @export
 #'
 #' @examples
 #' parameter_values(data_set = "fimaNfipClaims", data_field = "totalBuildingInsuranceCoverage")
-parameter_values <- memoise(function(data_set = NULL, data_field = NULL) {
+parameter_values <- memoise::memoise(function(data_set = NULL, data_field = NULL) {
   data <- open_fema(data_set = data_set, top_n = 999, select = data_field, ask_before_call = F)
   values <- unique(data[, data_field])
 
