### 2021-11-15 #484 @dylan-turner25
@emilyriederer , thank you for the initial comments. The latest commit for the package has attempted to address each of these.
