+++ NEWS.md
@@ -1,0 +1,32 @@
+
+rfema 0.0.0.9000 (2021-11-19)
+=========================
+
+### DOCUMENTATION FIXES
+* Added a NEWS.md file 
+
+<!-- ### NEW FEATURES -->
+
+<!--   * New function added `do_things()` to do things (#5) -->
+
+<!-- ### MINOR IMPROVEMENTS -->
+
+<!--   * Improved documentation for `things()` (#4) -->
+
+<!-- ### BUG FIXES -->
+
+<!--   * Fix parsing bug in `stuff()` (#3) -->
+
+<!-- ### DEPRECATED AND DEFUNCT -->
+
+<!--   * `hello_world()` now deprecated and will be removed in a -->
+<!--      future version, use `hello_mars()` -->
+
+<!-- ### DOCUMENTATION FIXES -->
+
+<!--   * Adding a NEWS.md file -->
+
+<!-- ### (a special: any heading grouping a large number of changes under one thing) -->
+
+<!--     * blablabla. -->
+
