### 2021-12-05 #484 @emilyriederer
Hi @fawda123 and @fmichonneau ! I hope you both are doing well. This is just a friendly reminder that the target review deadline for `rfema` is about a week away. Let us know if you have any questions or concerns. Otherwise, looking forward to seeing all of your feedback!
