### 2021-12-22 ebfdc04
add comment showing code made uncessary with rfema
