### 2021-12-22 ffbe5c4
replace devtools with remotes for GH installation
