+++ R/bulk_dl.R
@@ -46,7 +46,7 @@ bulk_dl <- function(data_set, output_dir = NULL, file_name = NULL, size_warning 
     file_name <- paste0(data_set, ".csv")
   }
 
-  if (size_warning == T) {
+  if (size_warning) {
     file_size <- ds$distribution.datasetSize
     if (grepl("large", file_size)) {
       message(paste0(
@@ -58,7 +58,7 @@ bulk_dl <- function(data_set, output_dir = NULL, file_name = NULL, size_warning 
       user_response <- readline(prompt = " 1 - Yes, get that data!, 0 - No:")
 
       if (user_response != "1") {
-        stop("Opperation aborted by user.")
+        stop("Operation aborted by user.")
       }
     }
   }
