### 2021-12-22 29843c8
download in binary for all oses
the `mode = "wb"` is important on windows, but as far as I understand, there are no downsides to using it on other OSes, so you can use it regardless everywhere else.
