+++ R/bulk_dl.R
@@ -64,18 +64,9 @@ bulk_dl <- function(data_set, output_dir = NULL, file_name = NULL, size_warning 
   }
 
   message(paste0("Downloading file to ", paste0(output_dir, "/", file_name)))
-
-  # identify opperating system
-  system <- tolower(Sys.info())
-  if (T %in% grepl("windows", system)) {
-    utils::download.file(url,
+  
+  utils::download.file(url,
       destfile = paste0(output_dir, "/", file_name),
       mode = "wb"
     )
-  }
-  if (T %in% grepl("linux", system)) {
-    utils::download.file(url, destfile = paste0(output_dir, "/", file_name))
-  } else {
-    utils::download.file(url, destfile = paste0(output_dir, "/", file_name))
-  }
 }
