### 2021-12-23 #484 @fmichonneau
The need for the package is well justified. When reading quickly the README the first time, I was slightly confused by the first block of code that demonstrates what it would be to not have the package to work with this data. I suggest adding a comment in the block of text to clarify this: https://github.com/dylan-turner25/rfema/pull/3

The installation instructions are straightforward. I suggest you recommend your users to use the `remotes` package instead of `devtools` to install your package from GitHub: https://github.com/dylan-turner25/rfema/pull/4

The vignette is clear and comprehensive. However, the source of the vignette file is in the `.Rbuildignore` file which means that it's not included in the package/available from the R console (it is however available from the pkgdown generated site). If your intention is to pre-compute the vignette so it's not generated when the package gets built, you can refer to [Joeren's blog post](https://ropensci.org/blog/2019/12/08/precompute-vignettes/).

It's only something to worry about after the approval, but it'd be useful to add the URL for the pkgdown-generated site in the GitHub repo info.

I also notice that `T` or `F` is used instead of `TRUE` or `FALSE` in the function declarations. I encourage you to use the long-form in the documentation.

If you plan to submit your package to CRAN, I recommend you wrap your examples in your function help pages with `\dontrun{}` to avoid failures in case the API is down when CRAN performs their daily checks.

The link to the code of conduct from the CONTRIBUTING file is broken.

---

Thanks for the opportunity to review your package @dylan-turner25. I really don't know much about this domain so it is difficult for me to come up with good questions to try to explore the package and the FEMA API on my own. However, your vignette and the examples give a good indication of how to use the package and interact with the API. I especially appreciated the extra examples you provide in the vignette. I think the package's interface gives the appropriate building blocks to interact with an API that provides access to a lot of relatively complex data.

I agree with everything that @fawda123 said in his excellent and comprehensive review (I read his review only after interacting with the package for a little bit). He expressed elequantly the main things I found about the package so I'm not going to repeat them. I do encourage you to proofread carefully your README, vignette, and help files. The `devtools::spell_check()` function could help with that.

I noticed that you use `memoise` on your internal package functions. The memoise README notes that for packages, the memoization should be done in the `.onLoad` function (see the last paragraph of the ["Details" section](https://memoise.r-lib.org/reference/memoise.html#details)).

I submitted a few pull requests on your package for suggestions to improve/simplify the code. Don't hesitate to let me know if you have questions:
- https://github.com/dylan-turner25/rfema/pull/3
- https://github.com/dylan-turner25/rfema/pull/4
- https://github.com/dylan-turner25/rfema/pull/5
- https://github.com/dylan-turner25/rfema/pull/6

Great work on this package and on making public data easily accessible in the R ecosystem!
