### 2022-01-01 2adffd0
Working on suggestions from Ropensci reviewers
