+++ R/fema_data_fields.R
@@ -3,7 +3,7 @@
 #'
 #' @param data_set a character string indicating the data set of interest
 #'
-#' @return Returns a data frame consisting of the data fields name,
+#' @return Returns a tibble consisting of the data fields name,
 #' along with information about each data field including the data type,
 #' a description of the data field, and whether the data field is "searchable"
 #'  (i.e. can it be used to filter the returned data in an API query)
@@ -12,7 +12,7 @@
 #' @examples
 #' fema_data_fields("FimaNfipClaims")
 #' fema_data_fields("FimaNfipPolicies")
-fema_data_fields <- memoise::memoise(function(data_set) {
+fema_data_fields <- function(data_set) {
 
 
 
@@ -28,4 +28,4 @@ fema_data_fields <- memoise::memoise(function(data_set) {
 
 
   return(data_fields)
-})
+}
