+++ R/helpers.R
@@ -1,3 +1,54 @@
+
+#' Helper function to convert date columns
+#'
+#' @param data a tibble returned from open_fema()
+#'
+#' @return returns a tibble with the columns representing dates converted to 
+#' POSIXct format
+#'
+#' @examples
+#' data <- open_fema("fimanfipclaims",top_n = 10)
+#' data_with_dates <- rfema:::convert_dates(data)
+convert_dates <- function(data){
+  
+  # identify columns to convert to date objects based on if "date" is 
+  # in the column name
+  cols_to_convert <- colnames(data)[grepl("date",tolower(colnames(data)))]
+
+  # loop over each column identified above  
+  for(c in cols_to_convert){
+    
+    # initialize an object to store the converted column
+    converted_col <- NULL
+    
+    # wrap the following in a tryCatch in case a column was identified above
+    # that really shouldn't be converted to a date object
+    tryCatch({
+      # convert the tibble to a data frame
+      to_convert <- data.frame(data)[,c]
+      
+      # replace NULL values with NA
+      to_convert <- replace(to_convert, to_convert == "NULL", NA)
+      
+      # convert to a POSIXct object
+      converted_col <- as.POSIXct(to_convert, Sys.timezone())
+      
+    }, error = function(e){cat("ERROR :",conditionMessage(e), "\n")})
+    
+    # if the converted_col is not empty, go ahead and overwrite the 
+    # corresponding column in the data
+    if(is.null(converted_col) == F){
+      data[,c] <- converted_col
+    }
+    
+  }
+  
+  return(data)
+
+}
+
+
+
 #' Helper function to insure data_set parameter matches an open FEMA data
 #'  offering.
 #'
@@ -9,14 +60,12 @@
 #' @return returns the data with capitalization changed to be consistent with
 #' FEMA's naming convention or returns an error if the data set is not one of
 #' the valid offerings through the FEMA API.
-#' @export
 #' 
-#' @importFrom memoise memoise
 #'
 #' @examples
-#' valid_dataset("fimanfipclaims")
-#' valid_dataset("fIMANfipclaiMS")
-valid_dataset <- memoise::memoise(function(data_set) {
+#' rfema:::valid_dataset("fimanfipclaims")
+#' rfema:::valid_dataset("fIMANfipclaiMS")
+valid_dataset <- function(data_set) {
 
 
   # get df with info on fema data sets
@@ -38,7 +87,7 @@ valid_dataset <- memoise::memoise(function(data_set) {
     data_set <- match
   }
   return(as.character(unique(data_set)))
-})
+}
 
 
 
@@ -59,18 +108,15 @@ valid_dataset <- memoise::memoise(function(data_set) {
 #' @return The function returns a string containing a url that can be used to
 #'  query the API.
 #'  
-#' @importFrom memoise memoise
 #' 
-#' @export
 #'
 #' @examples
 #' filter_list <- list(baseFloodElevation = c(5, 6), countyCode = "34029")
 #' vars_to_select <- c("countyCode", "baseFloodElevation")
-#' url <- gen_api_query(
+#' url <- rfema:::gen_api_query(
 #'   data_set = "fimaNfipPolicies", top_n = 100,
-#'   filters = filter_list, select = vars_to_select
-#' )
-gen_api_query <- memoise::memoise(function(data_set = NULL, top_n = NULL, filters = NULL,
+#'   filters = filter_list, select = vars_to_select)
+gen_api_query <- function(data_set = NULL, top_n = NULL, filters = NULL,
                                   select = NULL) {
 
   # replace top_n with 1000 if no value is supplied
@@ -90,7 +136,7 @@ gen_api_query <- memoise::memoise(function(data_set = NULL, top_n = NULL, filter
 
[…]
