### 2022-01-01 6f9ad34
Working on issues raised during rOpensci review
