### 2022-01-01 18963df
adding test coverage for convert_dates() function.
