### 2022-01-01 a190f14
adding test coverage for time_iterations() function
