+++ NEWS.md
@@ -1,6 +1,6 @@
 
 
-rfema 0.0.1.9000 (2021-12-28)
+rfema 1.0.0 (2021-12-28)
 =========================
 ### NEW FEATURES 
 
