+++ R/zzz.R
@@ -2,7 +2,7 @@
   
   # memoise functions
   open_fema <<- memoise::memoise(open_fema)
-  feam_data_fields <<- memoise::memoise(fema_data_fields)
+  fema_data_fields  <<- memoise::memoise(fema_data_fields)
   fema_data_sets <<- memoise::memoise(fema_data_sets)
   valid_dataset <<- memoise::memoise(valid_dataset)
   gen_api_query <<- memoise::memoise(gen_api_query)
