+++ man/get_status_message.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helpers.R
+\name{get_status_message}
+\alias{get_status_message}
+\title{Helper function to construct status message for \code{open_fema()} function}
+\usage{
+get_status_message(n_records, iterations, top_n, estimated_time)
+}
+\arguments{
+\item{n_records}{a numeric value indicating the number of matching records
+in the API call}
+
+\item{iterations}{numeric value indicating the  number of iterations
+necessary to retrieve all the data}
+
+\item{top_n}{numeric value indicating the number of matching records to return}
+
+\item{estimated_time}{character string indicating the estimated time to
+complete the API call.}
+}
+\value{
+a character string
+}
+\description{
+Helper function to construct status message for \code{open_fema()} function
+}
+\examples{
+rfema:::get_status_message(2000,2,1000,"2 minutes")
+}
