### 2022-01-05 1277675
wrapping examples in function documentaion in \dontrun{}
