+++ R/fema_data_fields.R
@@ -10,8 +10,8 @@
 #' @export
 #'
 #' @examples
-#' fema_data_fields("FimaNfipClaims")
-#' fema_data_fields("FimaNfipPolicies")
+#' \dontrun{fema_data_fields("FimaNfipClaims")}
+#' \dontrun{fema_data_fields("FimaNfipPolicies")}
 fema_data_fields <- function(data_set) {
 
 
