+++ R/valid_parameters.R
@@ -12,7 +12,7 @@
 #' @export
 #'
 #' @examples
-#' valid_parameters("fimaNfipPolicies")
+#' \dontrun{valid_parameters("fimaNfipPolicies")}
 valid_parameters <- memoise::memoise(function(data_set = NULL) {
   data_set <- valid_dataset(data_set = data_set)
   data_fields <- fema_data_fields(data_set)
