+++ man/gen_api_query.Rd
@@ -27,9 +27,9 @@ query the API.
 Helper function to generate the URL string for an API query
 }
 \examples{
-filter_list <- list(baseFloodElevation = c(5, 6), countyCode = "34029")
+\dontrun{filter_list <- list(baseFloodElevation = c(5, 6), countyCode = "34029")
 vars_to_select <- c("countyCode", "baseFloodElevation")
 url <- rfema:::gen_api_query(
   data_set = "fimaNfipPolicies", top_n = 100,
-  filters = filter_list, select = vars_to_select)
+  filters = filter_list, select = vars_to_select)}
 }

[…]
