### 2022-01-06 280ce5f
Updating documentation and precompiling vigette that depends on the FEMA API being up.
