+++ R/bulk_dl.R
@@ -1,6 +1,9 @@
-#' Bulk download full openFEMA data sets as .csv files
+#' Bulk download full open FEMA data sets as .csv files
 #'
-#' For large files, the function
+#' For large data requests, it can sometimes be more practical (and quicker) 
+#' to download and work with the entire data set. This function provides an 
+#' easy way to download any of FEMA's data sets in its entirety and save it 
+#' locally as a csv file. 
 #'
 #' @param data_set A character string indicating the name of the data set
 #' to download
@@ -17,7 +20,7 @@
 #' @examples
 #' \dontrun{bulk_dl("femaregions")} # download the file
 #' \dontrun{file.remove("FemaRegions.csv")} # clean up directory after file downloads
-bulk_dl <- function(data_set, output_dir = NULL, file_name = NULL, size_warning = T) {
+bulk_dl <- function(data_set, output_dir = NULL, file_name = NULL, size_warning = TRUE) {
 
   # get info on all data sets fema offers
   ds <- fema_data_sets()
