+++ man/valid_dataset.Rd
@@ -17,7 +17,7 @@ the valid offerings through the FEMA API.
 }
 \description{
 The function corrects any capitalization inconsistencies so that the
-"data_set" parameter in any of rfema's functions is not case sensative.
+"data_set" parameter in any of rfema's functions is not case sensitive.
 }
 \examples{
 \dontrun{rfema:::valid_dataset("fimanfipclaims")}
