### 2022-01-08 afb6bc6
Updating documentation, reademe, and vignette
