+++ man/valid_dataset.Rd
@@ -1,25 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helpers.R
-\name{valid_dataset}
-\alias{valid_dataset}
-\title{Helper function to insure data_set parameter matches an open FEMA data
-offering.}
-\usage{
-valid_dataset(data_set)
-}
-\arguments{
-\item{data_set}{a character string representing the data set.}
-}
-\value{
-returns the data with capitalization changed to be consistent with
-FEMA's naming convention or returns an error if the data set is not one of
-the valid offerings through the FEMA API.
-}
-\description{
-The function corrects any capitalization inconsistencies so that the
-"data_set" parameter in any of rfema's functions is not case sensitive.
-}
-\examples{
-\dontrun{rfema:::valid_dataset("fimanfipclaims")}
-\dontrun{rfema:::valid_dataset("fIMANfipclaiMS")}
-}
