+++ man/time_iterations.Rd
@@ -1,23 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helpers.R
-\name{time_iterations}
-\alias{time_iterations}
-\title{Helper function to estimate time per API call}
-\usage{
-time_iterations(data_set, iterations)
-}
-\arguments{
-\item{data_set}{the name of the open FEMA data set to time API calls for}
-
-\item{iterations}{the number of iterations that need to be performed}
-}
-\value{
-returns a character string detailing the estimated time needed to
-perform the specified number of API call iterations on the specified data set
-}
-\description{
-Helper function to estimate time per API call
-}
-\examples{
-\dontrun{rfema:::time_iterations("fimanfipclaims",25)}
-}
