+++ man/convert_dates.Rd
@@ -1,22 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helpers.R
-\name{convert_dates}
-\alias{convert_dates}
-\title{Helper function to convert date columns}
-\usage{
-convert_dates(data)
-}
-\arguments{
-\item{data}{a tibble returned from open_fema()}
-}
-\value{
-returns a tibble with the columns representing dates converted to
-POSIXct format
-}
-\description{
-Helper function to convert date columns
-}
-\examples{
-\dontrun{data <- open_fema("fimanfipclaims",top_n = 10)}
-\dontrun{data_with_dates <- rfema:::convert_dates(data)}
-}
