### 2022-01-08 3ffd40e
adding vignette to inst/doc folder
