+++ vignettes/getting_started.Rmd
@@ -90,6 +90,7 @@ The FEMA API limits the number of records that can be returned in a single query
 
 ```
 #> Calculating estimated API call time...
+#> Error in curl::curl_fetch_memory(url, handle = handle): Operation was aborted by an application callback
 #> 2564279 matching records found. At 1000 records per call, it will take 2565 individual API calls to get all matching records. It's estimated that this will take approximately 2.01 hours. Continue?
 #> [1] 1 - Yes, get that data!, 0 - No, let me rethink my API call:
 ```
