+++ R/fema_data_sets.R
@@ -4,16 +4,18 @@
 #' @return Returns a tibble containing meta data about each data
 #' set available through the FEMA API. For more information see the FEMA
 #' documentation page: https://www.fema.gov/about/openfema/data-sets.
-#' 
+#'
 #' @import httr
 #' @importFrom plyr rbind.fill
 #' @importFrom tibble as_tibble
-#' 
+#'
 #' @export
 #'
 #' @examples
-#' \dontrun{fema_data_sets()}
-fema_data_sets <- function(){
+#' \dontrun{
+#' fema_data_sets()
+#' }
+fema_data_sets <- function() {
   result <- httr::GET("https://www.fema.gov/api/open/v1/DataSets")
   jsonData <- httr::content(result)
   df <- data.frame(t(unlist(jsonData$DataSets[1])))
