+++ tests/testthat/test_parameter_values.R
@@ -11,15 +11,13 @@ test_that(paste0("returned object is a tibble of dimension 1 x N"), {
     returned_obj <- parameter_values("fimanfipclaims", "floodZone", message = FALSE)
   })
   expect_equal(tibble::is_tibble(returned_obj), TRUE) # is a tibble
-  expect_equal(ncol(returned_obj) > 1,  TRUE) # has a single row
-  expect_equal(nrow(returned_obj) , 1) # has multiple columns
+  expect_equal(ncol(returned_obj) > 1, TRUE) # has a single row
+  expect_equal(nrow(returned_obj), 1) # has multiple columns
 })
 
 test_that(paste0("returned object is NULL when message argument is TRUE"), {
   vcr::use_cassette("parameter_value2", {
     returned_obj <- parameter_values("fimanfipclaims", "floodZone", message = TRUE)
   })
-  expect_equal(returned_obj, NULL) 
+  expect_equal(returned_obj, NULL)
 })
-
-
