+++ R/bulk_dl.R
@@ -1,9 +1,9 @@
 #' Bulk download full open FEMA data sets as .csv files
 #'
-#' For large data requests, it can sometimes be more practical (and quicker) 
-#' to download and work with the entire data set. This function provides an 
-#' easy way to download any of FEMA's data sets in its entirety and save it 
-#' locally as a csv file. 
+#' For large data requests, it can sometimes be more practical (and quicker)
+#' to download and work with the entire data set. This function provides an
+#' easy way to download any of FEMA's data sets in its entirety and save it
+#' locally as a csv file.
 #'
 #' @param data_set A character string indicating the name of the data set
 #' to download
@@ -18,8 +18,12 @@
 #' @importFrom utils download.file
 #' @export
 #' @examples
-#' \dontrun{bulk_dl("femaregions")} # download the file
-#' \dontrun{file.remove("FemaRegions.csv")} # clean up directory after file downloads
+#' \dontrun{
+#' bulk_dl("femaregions") # download the file
+#' }
+#' \dontrun{
+#' file.remove("FemaRegions.csv") # clean up directory after file downloads
+#' }
 bulk_dl <- function(data_set, output_dir = NULL, file_name = NULL, size_warning = TRUE) {
 
   # get info on all data sets fema offers
@@ -67,9 +71,9 @@ bulk_dl <- function(data_set, output_dir = NULL, file_name = NULL, size_warning 
   }
 
   message(paste0("Downloading file to ", paste0(output_dir, "/", file_name)))
-  
+
   utils::download.file(url,
-      destfile = paste0(output_dir, "/", file_name),
-      mode = "wb"
-    )
+    destfile = paste0(output_dir, "/", file_name),
+    mode = "wb"
+  )
 }
