+++ tests/testthat/test_valid_dataset.R
@@ -14,8 +14,7 @@ test_that("characters of input and output match", {
 })
 
 
-test_that("error is returned when invalid data set argument is used",{
- err <- expect_error(valid_dataset(data_set = "asdfasdfa"))
- expect_true(grepl("is not a valid data set", err$message))
+test_that("error is returned when invalid data set argument is used", {
+  err <- expect_error(valid_dataset(data_set = "asdfasdfa"))
+  expect_true(grepl("is not a valid data set", err$message))
 })
-
