### 2022-01-09 03b7454
adding .md version of vignette for viewing directly on github
