### 2022-01-09 0bfd9c4
removing header from getting_started.md
