+++ vignettes/getting_started.md
@@ -1,13 +1,3 @@
----
-title: "rfema: Getting Started"
-author: "Dylan Turner"
-date: "2022-01-08"
-output:
- github_document:
-   toc: true
-   toc_depth: 3
----
-
 
 
 ## Introduction
@@ -85,7 +75,7 @@ The FEMA API limits the number of records that can be returned in a single query
 
 ```
 #> Calculating estimated API call time...
-#> 2564279 matching records found. At 1000 records per call, it will take 2565 individual API calls to get all matching records. It's estimated that this will take approximately 2.1 hours. Continue?
+#> 2564279 matching records found. At 1000 records per call, it will take 2565 individual API calls to get all matching records. It's estimated that this will take approximately 3.44 hours. Continue?
 #> [1] 1 - Yes, get that data!, 0 - No, let me rethink my API call:
 ```
 
