+++ vignettes/github/getting_started.Rmd
@@ -1,519 +1,0 @@
----
-title: "rfema: Getting Started"
-author: "Dylan Turner"
-date: "2022-01-08"
-output:
- github_document:
-   toc: true
-   toc_depth: 3
----
-
-
-
-## Introduction
-
-This vignette provides a brief overview on using the `rfema` package to obtain data from the Open FEMA API. The rest of this vignette covers how to install the package, followed by examples on using the package to obtain data for various objectives. 
-
-## Installation
-Right now, the best way to install and use the `rfema` package is by installing directly from GitHub using `remotes::install_github("dylan-turner25/rfema")`. The FEMA API does not require and API key, meaning no further setup steps need be taken to start using the package
-
-## Avaliable Datasets
-For those unfamiliar with the data sets available through the FEMA API, a good starting place is to visit the [FEMA API documentation page](https://www.fema.gov/about/openfema/data-sets). However, if you are already familiar with the data and want to quickly reference the data set names or another piece of meta data, using the `fema_data_sets()` function to obtain a tibble of available data sets along with associated meta data is a convenient option.
-
-```r
-# store the avaliable data sets as an object in your R environment that can be referenced later
-data_sets <- fema_data_sets() 
-
-# view data 
-data_sets
-#> # A tibble: 37 × 64
-#>    identifier name  title description distribution.ac… distribution.fo… distribution.da… distribution.ac… distribution.fo… distribution.da… distribution.ac…
-#>    <chr>      <chr> <chr> <chr>       <chr>            <chr>            <chr>            <chr>            <chr>            <chr>            <chr>           
-#>  1 openfema-1 Publ… Publ… FEMA provi… https://www.fem… csv              small (10MB - 5… https://www.fem… json             medium (50MB - … https://www.fem…
-#>  2 openfema-1 Publ… Publ… FEMA provi… https://www.fem… csv              small (10MB - 5… https://www.fem… json             medium (50MB - … https://www.fem…
-#>  3 openfema-… Fema… FEMA… This data … https://www.fem… csv              medium (50MB - … https://www.fem… json             medium (50MB - … https://www.fem…
-#>  4 openfema-… Haza… Haza… This datas… https://www.fem… csv              small (10MB - 5… https://www.fem… json             small (10MB - 5… https://www.fem…
-#>  5 openfema-… Fema… FEMA… Provides t… https://www.fem… csv              tiny (< 10 MB)   https://www.fem… json             tiny (< 10 MB)   https://www.fem…
-#>  6 openfema-… Emer… Emer… This datas… https://www.fem… csv              tiny (< 10MB)    https://www.fem… json             tiny (< 10MB)    https://www.fem…
-#>  7 openfema-… Haza… Haza… The datase… https://www.fem… csv              small (10MB - 5… https://www.fem… json             small (10MB - 5… https://www.fem…
-#>  8 openfema-… Hous… Hous… This datas… https://www.fem… csv              small (10MB - 5… https://www.fem… json             medium (50MB - … https://www.fem…
-#>  9 openfema-8 Data… Open… Metadata f… https://www.fem… csv              tiny (< 10MB)    https://www.fem… json             tiny (< 10MB)    https://www.fem…
-#> 10 openfema-… Haza… Haza… This datas… https://www.fem… csv              small (10MB - 5… https://www.fem… json             medium (50MB - … https://www.fem…
-#> # … with 27 more rows, and 53 more variables: distribution.format.2 <chr>, distribution.datasetSize.2 <chr>, webService <chr>, dataDictionary <chr>,
[…]
