### 2022-01-10 7ed146e
updating pkgdown site and adding vignettes to .Rbuildignore to avoid qpdf installation problem.
