### 2022-01-10 #484 @dylan-turner25
@emilyriederer, @fawda123 , and @fmichonneau, I have incorporated all the feedback I received and documented all the changes made in the NEWS.md file. I'll note that I'm still kind of perplexed on the best practices for making vignettes available.  My understanding is a package down site will document everything once the package is hosted with the other rOpensci packages. When I generate the package down site for this package, everything seems to be working as intended with all documentation and vignettes made available. I also made a copy of my vignette as a .md file so it can be viewed directly on Github (the README links directly to this .md file). Let me know if some other method of vignette distribution is a better choice.

Thanks again for all your feedback,
Dylan
