### 2022-01-10 #484 @fmichonneau
Generally, you write your vignette as an Rmd file in the `vignettes/` directory.

When you build your package (using `R CMD build` from your terminal or `devtools::build()` from your R console), then R will take your vignette, and put the rendered version in the archive file. When your users are going to install your package, this rendered vignette will be available to them. Currently, you have the path to your vignette in your `.Rbuildignore` and so this file is excluded at build time and the vignette is not available to your users if they try to locate the vignette directly on their systems using `browseVignettes("rfema")`. The vignette will also not be listed on CRAN if you plan to submit your package there.

The `pkgdown` package doesn't look at `.Rbuildignore` file and therefore builds the vignette and makes it available on the generated site. That's why you see it there.

If you leave it as it is, your vignette would only be available from the `pkgdown` site associated with your package but will not be included in the R package that your users would install on their system.
