### 2022-01-10 993465c
removing getting_started.Rmd from .Rbuildignore
