### 2022-01-10 #484 @dylan-turner25
@fmichonneau , thank you for the explanation. I just pushed a new commit that removed the path to my vignette's  .Rmd file from `.Rbuildignore`
