### 2022-01-16 #484 @emilyriederer
Hey @dylan-turner25 - thanks for working through the review feedback! @fmichonneau and @fawda123 , if you could please take on final look and confirm with the [approval template](https://devguide.ropensci.org/approval2template.html), that would be great!
