### 2022-01-23 #484 @fawda123
Hi there, apologies for the delay.  The revisions look fantastic, I especially like the addition of the `time_iterations()` function that estimates the API request time.  Very useful changes.  I've included one more PR with some very minor edits to the README.

* https://github.com/dylan-turner25/rfema/pull/7

Everything looks good on my end, nice work @dylan-turner25!

## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 6
